+++ src/Message.hpp
@@ -11,8 +11,8 @@ namespace message {
   int write_data(void **buf, int *buf_len, const void *data, size_t len);
   int read_data(void **buf, int *buf_len, void *data, size_t len);
 
-  int write_int(void **buf, int *buf_len, int value);
-  int read_int(void **buf, int *buf_len, int *value);
+  int write_int(void **buf, int *buf_len, long long value);
+  int read_int(void **buf, int *buf_len, long long *value);
 
   int write_float(void **buf, int *buf_len, float value);
   int read_float(void **buf, int *buf_len, float *value);
